BBWPC: Calculator for BedrijfsBodemWaterPlan (BBWP)

The BBWP Calculator can be used to calculate the opportunities and suitability of agronomic measures to improve the sustainability of agriculture. It includes the framework of the BedrijfsBodemWaterPlan (BBWP) developed in the Netherlands (Ros et al., 2020), as well the Dutch Ecoregelingen Framework developed by Boerennatuur and associated agricultural farmers collectives. Both come together in the BBWP tool to support farmers in their soil and water management. The BedrijfsBodemWaterPlan is a tool for The Netherlands that recommends field specific agronomic measures focusing on the improvement of soil health, the reduction of nutrient losses to groundwater and surface water and the water retention properties. More information about the BedrijfsBodemWaterPlan can be found at www.bedrijfsbodemwaterplan.nl.

See also

Author

Maintainer: Gerard Ros gerard.ros@nmi-agro.nl

Authors:

Sven Verweij sven.verweij@nmi-agro.nl

Other contributors:

Job de Pater job.depater@nmi-agro.nl [contributor]

Yuki Fujita yuki.fujita@nmi-agro.nl [contributor]

Kees van den Dool kees.vandendool@nmi-agro.nl [contributor]

Astrid Berndsen astrid.berndsen@nmi-agro.nl [contributor]

Brent Riechelman brent.riechelman@nmi-agro.nl (ORCID) [contributor]

Nutriënten Managment Instituut nmi@nmi-agro.nl [copyright holder]