An overview of correction factors per economic agricultural area

This table contains the correction factor for the financial reward differentiated per Agricultural Economic Region

Usage

er_aer_reward

Format

An object of class data.table (inherits from data.frame) with 14 rows and 3 columns.

Details

statcode

The unique code of the Agricultural Economic Region

statname

The name of the Agricultural Economic Region

er_cf

The value of the correction factor for the financial reward differentiated per Agricutural Economic Region