The importance and scoring of environmental challenges for Ecoregelingen

This table contains the objective scoring and correction factors to estimate score for Ecoregeling

Usage

er_scoring

Format

An object of class data.table (inherits from data.frame) with 8 rows and 7 columns.

Details

soiltype

Agricultural soil type

type

The type of weighting factor: urgency (distribution of scores depending on political urgency) or aim (distribution of scores depending on aim for environmental challenges)

cf_soil

Weighting factor for soil

cf_water

Weighting factor for water

cf_climate

Weighting factor for climate

cf_biodiversity

Weighting factor for biodiversity

cf_landscape

Weighting factor for landscape