Create the LSW data.table from regional soil properties per Local Surface Water or any Regional Geometry

For each of the fields (or spatial geometries) for which BBWP field properties are estimated, the function requires regional soil properties (mean and standard deviation). These properties can be derived from any spatial dataset. This helper function creates a data.table allowing unit checks on input variables.

Usage

bbwp_lsw_properties(
  B_LSW_ID,
  B_SOM_LOI = 8.65,
  B_CLAY_MI = 15.8,
  B_SAND_MI = 60.5,
  B_SILT_MI = 23.71,
  B_N_RT = 3834,
  B_P_AL = 49,
  B_P_CC = 2.71,
  B_P_WA = 40,
  B_P_SG = 22,
  B_FE_OX = 83,
  B_AL_OX = 40,
  B_RO_R = 0.5,
  B_SA_W = 0.47,
  B_SOM_LOI_SD = 6.67,
  B_CLAY_MI_SD = 13.45,
  B_SAND_MI_SD = 23.5,
  B_SILT_MI_SD = 11.7,
  B_N_RT_SD = 2928,
  B_P_AL_SD = 13.5,
  B_P_CC_SD = 1.51,
  B_P_WA_SD = 15.6,
  B_P_SG_SD = 14,
  B_FE_OX_SD = 59,
  B_AL_OX_SD = 19,
  B_RO_R_SD = 0.3,
  B_SA_W_SD = 0.33
)

Arguments

B_LSW_ID

(integer) An unique identifier for each Local Surface Water

B_SOM_LOI

(numeric) The mean percentage organic matter in the soil per LSW (%)

B_CLAY_MI

(numeric) The mean clay content of the soil per LSW (%)

B_SAND_MI

(numeric) The mean sand content of the soil per LSW (%)

B_SILT_MI

(numeric) The mean silt content of the soil per LSW (%)

B_N_RT

(numeric) The mean organic nitrogen content of the soil per LSW in mg N / kg

B_P_AL

(numeric) The mean P-AL content of the soil per LSW

B_P_CC

(numeric) The mean plant available P content, extracted with 0.01M CaCl2 per LSW (mg / kg)

B_P_WA

(numeric) The mean P-content of the soil extracted with water per LSW (mg P2O5 / 100 ml soil)

B_P_SG

(numeric) The mean P saturation degree per LSW (%)

B_FE_OX

(numeric) The mean aluminium content of soil per LSW (mmol / kg)

B_AL_OX

(numeric) The mean iron content of soil per LSW (mmol / kg)

B_RO_R

(numeric) The mean risk that surface water runs off the parcel per LSW

B_SA_W

(numeric) The mean wet perimeter index of all fields in a LSW, being the mean fraction of a field is surrounded by water

B_SOM_LOI_SD

(numeric) The standard deviation of percentage organic matter in the soil per LSW (%)

B_CLAY_MI_SD

(numeric) The standard deviation of clay content of the soil per LSW (%)

B_SAND_MI_SD

(numeric) The standard deviation of sand content of the soil per LSW (%)

B_SILT_MI_SD

(numeric) The standard deviation of silt content of the soil per LSW (%)

B_N_RT_SD

(numeric) The standard deviation of organic nitrogen content of the soil per LSW in mg N / kg

B_P_AL_SD

(numeric) The standard deviation of P-AL content of the soil per LSW

B_P_CC_SD

(numeric) The standard deviation of plant available P content, extracted with 0.01M CaCl2 per LSW (mg / kg)

B_P_WA_SD

(numeric) The standard deviation of P-content of the soil extracted with water per LSW (mg P2O5 / 100 ml soil)

B_P_SG_SD

(numeric) The standard deviation of P saturation degree per LSW (%)

B_FE_OX_SD

(numeric) The standard deviation of aluminium content of soil per LSW (mmol / kg)

B_AL_OX_SD

(numeric) The standard deviation of iron content of soil per LSW (mmol / kg)

B_RO_R_SD

(numeric) The standard deviation of the risk that surface water runs off the parcel per LSW

B_SA_W_SD

(numeric) The standard deviation of the wet perimeter index of all fields in a LSW

Details

Each user can apply its own estimates of regional soil properties (mean and sd) controlling fate of N and P in soils. If no information is given, the mean properties for all Dutch agricultural soils are used. The default spatial geometry is a Local surface Water, but any region can be used as long as the mean and sd of the relevant soil properties are known.